#feat: restrict deletion base on Ownership. Implemented a restriction to the deletion action, ensuring that only the owner of the task can request its deletion. If the user attempting to delete the task is not the owner, a Forbidden response (HTTP 403) is returned. This enhancement provides better security and prevents unauthorized access to task deletion.

+++ Source/HttpsRichardy.SimpleTask.Application/Commands/Handlers/DeleteTodoCommandHandler.cs
@@ -19,6 +19,9 @@ public class DeleteTodoCommandHandler : IRequestHandler<DeleteTodoCommand>
         if (todo is null)
             throw new ObjectDoesNotExistException($"the task with the ID '{request.TodoId}' does not exist.");
 
+        if (todo.UserId != request.UserId)
+            throw new UnauthorizedException();
+
         await _todoRepository.DeleteAsync(todo);
     }
 }
